feat: ✨ split study contacts page
